requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.1.11
django-cors-middleware==1.5.0
django-extensions==2.2.5
django-nose==1.4.6
django-oauth-toolkit==1.2.0
django-tenant-schemas==1.10.0
djangorestframework==3.10.2
djangorestframework-api-key==1.4.1
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nose==1.3.7
oauthlib==3.1.0
ordered-set==3.1.1
packaging==20.0
psycopg2-binary==2.8.4
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
uritemplate==3.0.1
urllib3==1.25.7
